[kdeutils-rules] Solve some of the problems with KRemoteControl's history.
KRemoteControl and KDELirc do not share an exactly linear
history. mzanetti even said the former could be considered a "fork" of
the latter.

Basically, at some points there were commits to kdelirc somewhere in
the tree, while kremotecontrol was also being worked on somewhere
else. However, the rules considered everything a single program, and
this was causing problems.

KDELirc's history is now treated separately from KRemoteControl's, in
that KDELirc's history starts when kdelirc is first imported into
/trunk/kdeutils and goes until it is moved to /tags/unmaintained/4,
and KRemoteControl originates from a copy of kdelirc to
/branches/work/kdelirc.