libkdeedu: Change qtmmlwidget and keduvoc to use separate branches.
I think it's better to keep it clear that eg. qtmmlwidget was *not* in
libkdeedu while it was in kalgebra: by putting the kalgebra-era code in a
separate branch with its own root, which gets merged into master later.

However, I use custom ref names (outside of 'refs/heads') so that it
doesn't actually create named git branches that would get pushed, but still
keeps a separate chain of commits.