kdeutils: Remove commented out rule creating `dead-master-2'.
The history here is a little bit different from the one which led to
`dead-master-1', as there was an `svn rm' in /trunk/KDE/kdeutils/ark/
which is only visible in the history of kdeutils itself, not Ark's.

To the latter, the code in trunk stays there until the code from
/branches/work/libarchive-based-ark was partly copied over it.

Thus kdeedu/parent-adder has been imported and a parentmap file has been
written for Ark. It replaces the billions of parents of r694732 with
r694571 (the last commit in libarchive-based-ark) and r672657 (the last
commit prior to Ark's removal from trunk in r692401).

Another script still needs to be written to remove the `old' directory
from trunk, as it was not copied from libarchive-based-ark in r694732.