translateRevlist: add support for disambiguating revisions using paths
You can now follow the revision number with an @ sign and a path, which
will be matched exactly against the SVN path in the commit message. This
is useful when a single SVN commit is converted into multiple git commits
based on paths. For example, to match a commit containing the following:

svn path=/branches/kstars/hdevalence/kstars/; revision=1121256

...you can use "1121256@/branches/kstars/hdevalence/kstars/".