kdeutils: Add a filter-branch script for superkaramba.
The original history is a bit messy, as superkaramba is created in
trunk only late in history (it starts in kdereview and then moves to
branches/KDE/3.5), so we artificially make it inherit from a previous
revision in the 3.5 branch.

Furthermore, git adds a weird parent to r457717, which we correct by
setting the previous commit as the single parent of it.