kstars: use a different approach to delete kstars/data/deepstars.dat.
I was using a svn2git rule deleting it only from trunk,
but it turns out the file was in the 'summer' branch too.
In addition, there was nothing left in the commit message to mention
the file got deleted.

I'm now using the good old bin/deleter-fb script, which deletes the file
via filter-branch *and* leaves a note in relevant commit messages.