Make all filterbranchish scripts take a $RULESETDIR env var, not $bindir.
This is consistent with some kdeutils scripts that already take RULESETDIR,
and lets scripts easily access files outside of bin (such as parent maps).

I also simplified the way scripts check for the var presence and error out
if it's not set, by using ${var:?error message} instead of a manual
"test -z" conditional. I'm not sure if this is a bashism, but just in case,
I changed all shebangs from /bin/sh to /bin/bash.