Put some doc history into 'doc2', copy it back to doc/ with filter-branch.
Export all the documentation history from doc/kmymoney (as opposed to
kmymoney/doc) into a "doc2" directory. This avoids problems in the commit
where the move happened.

Then I move the contents of doc2/ back into doc/ by post-processing
the repository with a filter-branch script. This is a bit tricky;
I can't just use git mv since there are commits for all four combinations
of doc and doc2 existing or not, and if both exist, we have to merge their
contents (move the contents of doc2 into doc).