Add recompress-svgz-and-png.sh and nullgzip.py tools.
nullgzip.py converts a gzip file to a null-compressed gzip file having
a valid gzip header, but completely uncompressed contents. This is
useful to produce valid SVGZ files that git can see as plain text and
produce meaningful deltas for.

recompress-svgz-and-png.sh rewrites the history of a repository,
converting all SVGZ files to null compression and applying aggressive
compression to all PNG files, so that "optimize-graphics" commits don't
lead to multiple copies of the same file in the history only differing
by their compression.