new helper function fill_branch
some first commits for a branch contain only the changed files in that
branch. This makes the git revision appear as deleting all the rest of
the tree. Here we filter all revisions in the branch to contain the
rest of the files, taken from the master revision right before
branching.