kdemultimedia: turn into a more traditional layout of rule files.
The kdemultimedia rule files were intentionally missing the final
'match /' rule, so that they could run at the same time by passing all
the files to svn2git at once, followed by a kdemm-bottom-rules file
containing *only* the 'match /' rule.

The problem is that then the rule files by themselves are incomplete,
they won't work without the bottom-rules file. This is incompatible with
the automation scripts I use to run conversions (especially the 'convrun'
that runs them on dewey).

In addition, I fear it could cause problems with the conversions
themselves: if a path is matched by more than one rule, its presence
in the repos would depend on the order the rule files are given to svn2git.

This commit adds the "match /" rule to all the individual rule files
and removes kdemultimedia-bottom-rules.