Rename the MODULE variable to APP.
In almost all generic rule files, such as feature-branches, MODULE is
expected to be a full KDE (sub)module name such as kdemultimedia/kmix.
In the kdemm rules, MODULE is being used for the kdemm application name
alone, which is incompatible with those shared rules.

Here I rename MODULE to APP, so that in the future I can use the common
rule files, where I'd give MODULE the value those rules expect.