Use common-kde-ignores to ignore branch/tag renames and tag miscreations.
This gets notably gets rid of r409205 and r411963, which renamed branches
and tags (respectively) in SVN. In the conversion, these commits cause a
mess whose gitk graph is hard to describe in words :) And they delete all
documentation, so the tip of the tags are doc-less, and for example
r428941 in kscd re-adds the documentation rather than modifying it.

It also gets rid of bad tag creations that were later deleted and
re-created. So here I also delete the now-redundant ignores in
common-kde-tags-rules.