Fix rule order for oldkmix1.
kdemm-common already has a rule for /trunk/kdemultimedia/kmix/
on all revisions, so the rule putting code in the oldkmix1 branch
was *never* matching.

Solution is to include kdemm-common after this special rule.
Anything not matched by it (due to the 'max revision')
will get caught by the general rule in kdemm-common.