Don't export /branches/work/doc/juk/.
I had to use a post-processing script and an ugly parentmap entry to get
all the branch commits (all two of them) to look good, and the final
merge commit (r1143923) turned out to be an evil merge[1], making it
hard to follow the history of the screenshots modified along with the
merge.

In the end I decided to just ignore the work branch. It's a big change
in a single commit, followed by a trivial typo fix. It ends up being
clearer to pretend the merge commit (r1143923) did the changes, and we
don't really lose useful history.

[1] "An evil merge is a merge that introduces changes that do not appear
in any parent." -- gitglossary(7) manpage