Add kompare parentmap and postprocess script
The postprocess script is just a start. Eventually we should have all
necessary postprocessing in here (so far mostly in the wiki).

The rules for kompare still have to be updated to make sense with the
parent map, but I wanted to commit the script as it might be helpful as
a starting point for other repos.