Fix warning when applying kompare parentmaps
r282532 exists as part of both KDE/3.2 and master, so we have to
disambiguate which of the two git commits has to be considered.
Without this, translateRevlist gives a warning, and it's undefined
which of the two commits it will use. Though it happened to pick the
correct one in this particular case, apparently.