Use the common rules file for kommander too.
kommander used to be part of quanta too, but it was in a different part of
its directory hierarchy. kdewebdev-common-rules now accepts variables to
specify where in the quanta directory the code used to be, so that
kommander-rules can use it.

I verified that the conversion result is unaffected for all four
repositories after this change.