Add more locations of the trunk history.
It's a bit complicated, because it was copied from kdebase to extragear
(they forgot to delete from kdebase immediately), and people kept
committing to both for a month without noticing the duplication. When the
problem was noticed, the code was deleted from kdebase and work continued
on extragear. However it seems there are some changes done in kdebase
during that month that were *never* merged back to extragear, so I want to
preserve that branch.