Force svn2git to reload directory after the move to scripts/
In r414920 we stop matching a filtered version of kde-i18n/ and switch to
matching kde-i18n/scripts/. But since we're staying in the same branch,
svn-all-fast-export will not re-read the directory contents. Anything
currently in the git repository will stay unless there is an explicit
delete operation found in SVN. This means files that were being matched
before (like README) but aren't present in SVN scripts/ at all will stay
in git, which doesn't match SVN history.

To work around this, we'll export the kde-i18n/ history into a separate
temporary ref (in refs/backups/ just so it doesn't appear as a real
branch). When we start matching kde-i18n/scripts/ on the 'master' branch,
the whole directory will be read again and files like README will not be
present, but it will still be marked as a child of the previous commit
because files were copied from kde-i18n/ to kde-i18n/scripts/.

In git it will appear as README etc. being *deleted* at r414920.